Rotation in all three axes (X, Y, and Z)

+++ drawing/vector3.go
@@ -1,19 +1,19 @@
 package drawing
 
 type Vector3 struct {
-	X int
-	Y int
-	Z int
+	X float64
+	Y float64
+	Z float64
 }
 
 func (vec *Vector3) MultiplyInt(n int) {
-	vec.X = vec.X * n
-	vec.Y = vec.Y * n
-	vec.Z = vec.Z * n
+	vec.X = vec.X * float64(n)
+	vec.Y = vec.Y * float64(n)
+	vec.Z = vec.Z * float64(n)
 }
 
 func (vec *Vector3) MultiplyFloat(n float64) {
-	vec.X = int(float64(vec.X) * n)
-	vec.Y = int(float64(vec.Y) * n)
-	vec.Z = int(float64(vec.Z) * n)
+	vec.X = vec.X * n
+	vec.Y = vec.Y * n
+	vec.Z = vec.Z * n
 }
